+++ Elements/src/Geometry/Transform.cs
@@ -493,7 +493,24 @@ namespace Elements.Geometry
         /// <param name="curve">The curve along which to orient the transform.</param>
         /// <param name="t">A parameter value between 0.0 and 1.0.</param>
         /// <param name="up"></param>
+
+        [Obsolete("Use CreateHorizontalFrameAlongCurve instead.")]
         public static Transform CreateOrientedAlongCurve(Curve curve, double t, Vector3 up = default(Vector3))
+        {
+            return CreateHorizontalFrameAlongCurve(curve, t, up);
+        }
+
+
+        /// <summary>
+        /// Create a transform that is oriented along 
+        /// a curve at parameter t. The transform's +z axis will align with 
+        /// the +z world axis, and the +x axis will align with the tangent
+        /// of the curve.
+        /// </summary>
+        /// <param name="curve">The curve along which to orient the transform.</param>
+        /// <param name="t">A parameter value between 0.0 and 1.0.</param>
+        /// <param name="up"></param>
+        public static Transform CreateHorizontalFrameAlongCurve(Curve curve, double t, Vector3 up = default(Vector3))
         {
             var temp = curve.TransformAt(t);
             return new Transform(temp.Origin, temp.ZAxis.Negate(), temp.XAxis.Negate(), Vector3.ZAxis);